move logo component
