+++ src/pages/namespace/Layout.tsx
@@ -14,7 +14,7 @@ import { Outlet, useParams } from "react-router-dom";
 import { useNamespace, useNamespaceActions } from "~/util/store/namespace";
 
 import Breadcrumb from "~/components/Breadcrumb";
-import Logo from "~/componentsNext/Logo";
+import Logo from "~/components/Logo";
 import { Menu } from "lucide-react";
 import Navigation from "~/components/Navigation";
 import NotificationMenu from "~/components/NotificationMenu";
